Fixed the decoder failed issue when decoding field-frame stream.
Description: blew cmd can't work no matter sys or video mem.
-c:v h264_qsv -i ~/Streams/Allegro_Cabac_CABAC_01_L30_SD576@25Hz_10.4.26l

Root cause: the ffmpeg parse the height is half of the MSDK's so the init
failed. Using the DecodeHeader of MSDK to replace the default parser and
make no change to the coded_width/height used as resoultion-change.

And make change the to surface allocator condition.

Bug:https://jira01.devtools.intel.com/browse/MN-2753